@@ -1,11 +1,14 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
-// import './index.css';
-// import App from './App';
+import ReactDOM from 'react-dom/client';
+import { BrowserRouter } from 'react-router-dom';
+import App from './App';
 
-// ReactDOM.render(
-//   <React.StrictMode>
-//     <App />
-//   </React.StrictMode>,
-//   document.getElementById('root')
-// );
+const root = ReactDOM.createRoot(
+    document.getElementById('root') as HTMLElement
+);
+
+root.render(
+    <BrowserRouter>
+        <App />
+    </BrowserRouter>
+);
