@@ -1,23 +1,30 @@
 import React from 'react';
+import { Route, Routes } from 'react-router-dom';
 
-// function App() {
-//   return (
-//     <div className="App">
-//       <header className="App-header">
-//         <p>
-//           Edit <code>src/App.tsx</code> and save to reload.
-//         </p>
-//         <a
-//           className="App-link"
-//           href="https://reactjs.org"
-//           target="_blank"
-//           rel="noopener noreferrer"
-//         >
-//           Learn React
-//         </a>
-//       </header>
-//     </div>
-//   );
-// }
+import './index.css';
 
-// export default App;
+import { initPage } from './api/main';
+import { Main } from './components/Main/Main';
+import { Header } from './components/Header/Header';
+import { Footer } from './components/Footer/Footer';
+import { Search } from './components/Search/Search';
+
+export function App() {
+    return (
+        <div className='app'>
+            <Routes>
+				<Route path="/" element={<>
+                    <Header refers="/search" ref_name="Поиск" />
+                    <Main onLoad={initPage} />
+                </>}></Route>
+                <Route path="search" element={<>
+                    <Header refers="/" ref_name="Главная" />
+                    <Search />
+                </>}></Route>
+			</Routes>
+            <Footer/>
+        </div>
+    );
+};
+
+export default App;
