@@ -1,17 +1,20 @@
 import { fetchArtistTopTrack, fetchTopArtists, fetchTopTracks } from './dataFetchers.js';
 
-const mostPlayableRoot = document.getElementById('most_playable_songs');
-const artistTopTrackRoot = document.getElementById('best_track_of_top_artist');
-const mostPopularArtistRoot = document.getElementById('most_popular_artists');
+let mostPlayableRoot = undefined;
+let artistTopTrackRoot = undefined;
+let mostPopularArtistRoot = undefined;
 
-const imageIndex = 2; // индекс разрешений картинок, полученных из api. 1 = medium
-
-window.onload = initPage;
+let imageIndex = undefined; // индекс разрешений картинок, полученных из api.
 
 /**
  * Функция инициализации страницы
  */
-function initPage() {
+export function initPage() {
+    mostPlayableRoot = document.getElementById('most_playable_songs');
+    artistTopTrackRoot = document.getElementById('best_track_of_top_artist');
+    mostPopularArtistRoot = document.getElementById('most_popular_artists');
+    imageIndex = 2;
+
     const topArtists = [];
     const topTracks = [];
     const topArtistTopTrack = [];
@@ -20,15 +23,14 @@ function initPage() {
         fetchTopArtists(topArtists),
         fetchTopTracks(topTracks)
     ])
-    .finally(() => {
-        fetchArtistTopTrack(topArtists[0].mbid, topArtistTopTrack)
-            .finally(() => {
-                handleMostPlayableTracks(topTracks);
-                handlePopularArtists(topArtists);
-                handleTopArtistTopTrack(topArtistTopTrack[0]);
-            })
-        }
-    );
+        .finally(() => {
+            fetchArtistTopTrack(topArtists[0].mbid, topArtistTopTrack)
+                .finally(() => {
+                    handleMostPlayableTracks(topTracks);
+                    handlePopularArtists(topArtists);
+                    handleTopArtistTopTrack(topArtistTopTrack[0]);
+                })
+        });
 }
 
 /**
@@ -50,7 +52,6 @@ export function handlePopularArtists(artists) {
             </a>`
         );
     }
-
 }
 
 /**
